clang: Upgrade to 16.0.5 version
Previously we are using bbappend for the clang recipes, however it is now
not applicable as there are changes for some directory structure in
llvm-project (Support/Triple.cpp to TargetParser/Triple.cpp). Hence,
clang 16.0.5 recipes are brought into meta-starfive instead.

Besides, for clang-native in clang_16.0.5, it is now not bootstrapped to be
aligned with our toolchain compilation steps. Our custom parameters are
currently separated with the bb file for easier maintanence going forward.

Co-authored by: Kai Zhe Mooi <kaizhe.mooi@starfivetech.com>
Signed-off-by: Jun Yuan Tan <junyuan.tan@starfivetech.com>