starfive-visionfive2: Fix PREFERRED_VERSION in machine config
PREFERRED_VERSION for linux-starfive-dev is change so that it is more generic
and do not need any changes when new commit is introduced. For GCC, Binutils
and glibc, they are changed to the generic variable which covers all of the
variants (e.g. crosssdk, native etc.). GCC version is updated as well as
upstream kirkstone gcc version is updated.

Signed-off-by: Jun Yuan Tan <junyuan.tan@starfivetech.com>